ASoC: topology: Fix a signedness bug in soc_tplg_dapm_widget_create()
BugLink: https://bugs.launchpad.net/bugs/1853382

[ Upstream commit 752c938a5c14b8cbf0ed3ffbfa637fb166255c3f ]

The "template.id" variable is an enum and in this context GCC will
treat it as an unsigned int so it can never be less than zero.

Fixes: 8a9782346dcc ("ASoC: topology: Add topology core")
Link: https://lore.kernel.org/r/20190925110624.GR3264@mwanda

+++ sound/soc/soc-topology.c
@@ -1472,7 +1472,7 @@ static int soc_tplg_dapm_widget_create(struct soc_tplg *tplg,
 
 	/* map user to kernel widget ID */
 	template.id = get_widget_id(w->id);
-	if (template.id < 0)
+	if ((int)template.id < 0)
 		return template.id;
 
 	/* strings are allocated here, but used and freed by the widget */
